+++ src/Model/Table/BannersTable.php
@@ -14,8 +14,8 @@ declare(strict_types=1);
 namespace MeCms\Model\Table;
 
 use ArrayObject;
+use Cake\Datasource\EntityInterface;
 use Cake\Event\Event;
-use Cake\ORM\Entity;
 use Cake\ORM\Query;
 use Cake\ORM\RulesChecker;
 use MeCms\Model\Table\AppTable;
@@ -36,13 +36,13 @@ class BannersTable extends AppTable
     /**
      * Called after an entity has been deleted
      * @param \Cake\Event\Event $event Event object
-     * @param \Cake\ORM\Entity $entity Entity object
+     * @param Cake\Datasource\EntityInterface $entity Entity object
      * @param \ArrayObject $options Options
      * @return void
      */
-    public function afterDelete(Event $event, Entity $entity, ArrayObject $options): void
+    public function afterDelete(Event $event, EntityInterface $entity, ArrayObject $options): void
     {
-        @unlink(BANNERS . $entity->filename);
+        @unlink(BANNERS . $entity->get('filename'));
 
         parent::afterDelete($event, $entity, $options);
     }