+++ src/Model/Table/PhotosAlbumsTable.php
@@ -14,8 +14,8 @@ declare(strict_types=1);
 namespace MeCms\Model\Table;
 
 use ArrayObject;
+use Cake\Datasource\EntityInterface;
 use Cake\Event\Event;
-use Cake\ORM\Entity;
 use Cake\ORM\Query;
 use Cake\ORM\RulesChecker;
 use MeCms\Model\Table\AppTable;
@@ -35,13 +35,13 @@ class PhotosAlbumsTable extends AppTable
     /**
      * Called after an entity has been deleted
      * @param \Cake\Event\Event $event Event object
-     * @param \Cake\ORM\Entity $entity Entity object
+     * @param Cake\Datasource\EntityInterface $entity Entity object
      * @param \ArrayObject $options Options
      * @return void
      */
-    public function afterDelete(Event $event, Entity $entity, ArrayObject $options): void
+    public function afterDelete(Event $event, EntityInterface $entity, ArrayObject $options): void
     {
-        @rmdir($entity->path);
+        @rmdir($entity->get('path'));
 
         parent::afterDelete($event, $entity, $options);
     }
@@ -49,13 +49,13 @@ class PhotosAlbumsTable extends AppTable
     /**
      * Called after an entity is saved
      * @param \Cake\Event\Event $event Event object
-     * @param \Cake\ORM\Entity $entity Entity object
+     * @param Cake\Datasource\EntityInterface $entity Entity object
      * @param \ArrayObject $options Options
      * @return void
      */
-    public function afterSave(Event $event, Entity $entity, ArrayObject $options): void
+    public function afterSave(Event $event, EntityInterface $entity, ArrayObject $options): void
     {
-        @mkdir($entity->path, 0777, true);
+        @mkdir($entity->get('path'), 0777, true);
 
         parent::afterSave($event, $entity, $options);
     }