+++ tests/TestCase/Command/AddUserCommandTest.php
@@ -89,7 +89,7 @@ class AddUserCommandTest extends TestCase
         $this->assertErrorContains('Field `password_repeat`: passwords don\'t match');
 
         //Tries with no groups
-        $Users->Groups->deleteAll(['id >=' => '1']);
+        $Users->Groups->deleteAll(['id IS NOT' => null]);
         $this->exec('me_cms.add_user -v');
         $this->assertExitWithSuccess();
         $this->assertErrorContains('Before you can manage users, you have to create at least a user group');