+++ tests/TestCase/Utility/SitemapTest.php
@@ -101,7 +101,6 @@ class SitemapTest extends TestCase
 
         //Deletes all records
         $table->deleteAll(['id IS NOT' => null]);
-        Cache::clearAll();
         $this->assertEmpty(Sitemap::pages());
     }
 
@@ -156,7 +155,6 @@ class SitemapTest extends TestCase
 
         //Deletes all records
         $table->deleteAll(['id IS NOT' => null]);
-        Cache::clearAll();
         $this->assertEmpty(Sitemap::photos());
     }
 
@@ -234,7 +232,6 @@ class SitemapTest extends TestCase
 
         //Deletes all records
         $table->deleteAll(['id IS NOT' => null]);
-        Cache::clearAll();
         $this->assertEmpty(Sitemap::posts());
     }
 
@@ -284,7 +281,6 @@ class SitemapTest extends TestCase
 
         //Deletes all records
         $table->deleteAll(['id IS NOT' => null]);
-        Cache::clearAll();
         $this->assertEmpty(Sitemap::postsTags());
     }