+++ tests/TestCase/View/Cell/PostsWidgetsCellTest.php
@@ -121,8 +121,7 @@ class PostsWidgetsCellTest extends CellTestCase
         $this->assertEquals(2, Cache::read('widget_categories', $this->Table->getCacheName())->count());
 
         //With no posts
-        Cache::clearAll();
-        $this->Table->deleteAll(['id >=' => 1]);
+        $this->Table->deleteAll(['id IS NOT' => null]);
         $this->assertEmpty($this->Widget->widget($widget)->render());
         $this->assertEmpty($this->Widget->widget($widget, ['render' => 'list'])->render());
     }
@@ -203,8 +202,7 @@ class PostsWidgetsCellTest extends CellTestCase
         $this->assertEquals(2, Cache::read('widget_latest_2', $this->Table->getCacheName())->count());
 
         //With no posts
-        Cache::clearAll();
-        $this->Table->deleteAll(['id >=' => 1]);
+        $this->Table->deleteAll(['id IS NOT' => null]);
         $this->assertEmpty($this->Widget->widget($widget)->render());
     }
 
@@ -286,8 +284,7 @@ class PostsWidgetsCellTest extends CellTestCase
         }
 
         //With no posts
-        Cache::clearAll();
-        $this->Table->deleteAll(['id >=' => 1]);
+        $this->Table->deleteAll(['id IS NOT' => null]);
         $this->assertEmpty($this->Widget->widget($widget)->render());
         $this->assertEmpty($this->Widget->widget($widget, ['render' => 'list'])->render());
     }